requirement.txt
asgiref==3.2.3
Django==3.0.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
Pillow==7.0.0
pytz==2019.3
psycopg2-binary = 2.9.3